Project Finalized

+++ Assets/Scripts/Data/ValueObjects/LevelData.cs
@@ -14,10 +14,9 @@ namespace Data.ValueObjects
         }
     }
 
-
     [Serializable]
     public struct PoolData
     {
-        public int RequiredObjectCount;
+        public byte RequiredObjectCount;
     }
 }
